Fix: Now supporting reading of some properties without calling Finish function
There are some websites (opencaching.com) which do not follow the official
steps to read data from the device. After starting the function StartDirectoryListing,
the opencaching.com website polls the result variable DirectoryListingXml
for changes and does not use the method FinishDirectoryListing to check if the
reading has finished. Which resulted in an always empty property DirectoryListingXml.