Fix: Header bytes of gmaptz.img are now removed after download
While updating the timezone file, the device downloads a .rgn file and saves it to
gmaptz.img. The unit will not accept this file as update unless the first 60 header
bytes are removed from the file. Compare to information found here:
http://www.noeman.org/gsm/garmin-maps/257457-garmin-timezone-map-v10-gmaptz-img.html

Files are now saved with a .tmp extension during download to prevent an incomplete file
if the connection drops. The file is renamed after the complete download. The original
plugin keeps the file completely in memory until it is fully downloaded.